convert condensed to shorthand...
I came up with this pattern in the car.

example.com/?dates=2018-01-01[1]a3b3c10a1

would mean starting with Jan 2018, the first day (a) gets slot 3 for show 1. The second day (b) would get slot 3, third day (c) slot 1. The zero denotes that we move to the next month. So february 1st gets slot 1.

a-E is 31 letters, so we have 31 days.

slots are 1-7 and each indicates a combination of morning, afternoon, and evening. (we don't have a number for none, since none would not even have a date).